requirements.txt
python-dotenv>=1.0.0
pytest>=8.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0